Oracle accepte finalement que ses mot-clés soient utilisés comme nom de colonne si on les met entre apostrophes (mais celles-ci sont refusées quand elles sont inutiles, fallait le savoir). Ca pose des problèmes pour écrire de Trigger, mais on n'a pas besoin du moins dans l'immédiat (et il semblerait qu'on puisse ruser en passant par une vue si ça devient nécessaire).
En conséquence, retour du champ {((mode}}} en rétablissant les fichiers pre[14588]. La MAJ pour les installations ayant chargé cette version restaurera ce nom de champ, les autres auront une mise à jour qui se réduira à changer le numéro de version installée, pas la peine de faire et défaire.


git-svn-id: svn://trac.rezo.net/spip/spip@14598 caf5f3e8-d4fe-0310-bb3e-c32d5e47d55d

+++ ecrire/maj/svn10000.php
@@ -228,9 +228,23 @@ $GLOBALS['maj'][14558] = array(array('upgrade_types_documents'));
 // etre sur qu'ils sont bien unipotents(?)...
 $GLOBALS['maj'][14559] = $GLOBALS['maj'][13904]+$GLOBALS['maj'][13929]+$GLOBALS['maj'][14558];
 
-$GLOBALS['maj'][14588] = array(
+
+// La version 14588 etait une mauvaise piste:
+// Retour en arriere pour ceux qui l'ont subi, ne rien faire sinon
+if (@$GLOBALS['meta']['version_installee'] >= 14588) {
+
+	// "mode" est un mot-cle d'Oracle
+	$GLOBALS['maj'][14588] = array(
 	array('sql_alter',"TABLE spip_documents  DROP INDEX mode"),
 	array('sql_alter',"TABLE spip_documents  CHANGE mode genre ENUM('vignette', 'image', 'document') DEFAULT 'document' NOT NULL"),
 	array('sql_alter',"TABLE spip_documents  ADD INDEX genre(genre)")
 			       );
+	// solution moins intrusive au pb de mot-clé d'Oracle, retour avant 14588
+	$GLOBALS['maj'][14598] = array(
+	array('sql_alter',"TABLE spip_documents  DROP INDEX genre"),
+	array('sql_alter',"TABLE spip_documents  CHANGE genre mode ENUM('vignette', 'image', 'document') DEFAULT 'document' NOT NULL"),
+	array('sql_alter',"TABLE spip_documents  ADD INDEX mode(mode)")
+			       );
+}
+
 ?>
